+++ src/app/shared/GroupNotes.svelte
@@ -75,6 +75,7 @@
       {feed}
       eager
       shouldListen
+      includeReposts
       contextAddress={address}
       skipNetwork={isGroupAddress(address)} />
   {/key}